refactor(project): ui와 domain 패키지로 분리한다
