+++ src/main/java/lotto/ui/InputView.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.ui;
 
 import java.util.Arrays;
 import java.util.List;
